[FEATURE] Enable INCLUDE_TYPOSCRIPT for be_layouts
The backend layout config field contains a TypoScript like syntax
to define the columns and rows. After all TypoScript containing fields
(setup, constants, userTS, pageTs) understand the INCLUDE_TYPOSCRIPT
notation to read the configuration from an external file, this should
be possible for the backend layout config field as well.

The patch only enables the reading out of files, the whole write to
file functionality is not provided.

Resolves: #28880
Releases: 6.0, 6.1

Change-Id: Ic43d103cf64b904c62e0866a252bb529f9091bb2