[BUGFIX] Temporary DB treemount notice missing in browser.php
If "Stop Page Tree" is set for a page, it is possible to set temporary
DB treemounts in the element browser/link browser. If such a treemount
is set, a notice is displayed right above the tree.

This notice is not present when the link browser "browser.php" is used
(e.g. page links in a FlexForm). Therefore a previously set mount cannot
be cancelled.

Change-Id: I942aad0e54d17ceb793008850f0563bb416503be
Releases: 6.1, 6.0, 4.7, 4.6, 4.5
Fixes: #43885