[BUGFIX] Update checksums after file processing
When recreating an already processed file the LocalImageProcessor
updates the width, height and filesize of the new image but not
the checksum.

Because of that the file will be deleted in FileProcessingService
when $processedFile->isProcessed() is called. Since the checksum
is wrong this call assumes that the file needs to be recreated and
deletes the "old" file on the go.

This will lead to an exception in this request, when getPublicUrl is
called on this object.

Fixes: #44073
Releases: 6.1, 6.0

Change-Id: Ib2df87b54866262e0092fbb24b4222f0b298e9d6
Reviewed-on: https://review.typo3.org/17207
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel
Reviewed-on: https://review.typo3.org/17208