[BUGFIX] Respect ignoreRootLevelRestriction in getExcludeFields
For TYPO3 6.0 a new security configuration

$TCA[$table]['security']['ignoreRootLevelRestriction']

was introduced together with the introduction of FAL. Before it was not
possible for a normal user to edit records with pid 0, but because FAL
saves sys_file and sys_file_reference records on pid 0, a new solution
was needed.

This must also be respected by
\TYPO3\CMS\Backend\Utility\BackendUtility:getExcludeFields
If not, it is not possible to define exclude fields for such tables
(e.g. field "link" in sys_file_reference).

Change-Id: I9cff7f4eaf8bff388781f1fcfddba0861449cb14
Releases: 6.1, 6.0
Fixes: #43849