[BUGFIX] Include file tables in access list
The BackendUtility::getExcludeFields method does not respect the
"ignoreRootLevelRestriction" configuration for these tables.
The result is that editors never see exclude fields as it is not
possible to set them in the access list in a backend group record.

Solution is to ignore the root level configuration if
"ignoreRootLevelRestriction" is also set in TCA.

This change also adds unit tests to check the fixed behavior
as well as the default behavior of this method.

Fixes: #44242
Releases: 6.0, 6.1

Change-Id: I2bb0d989028134388b24fa7ae4f139f2ebf8ee93