[BUGFIX] Fix email validation due to wrong use of IDNA
The IDNA converter is used to prepare for international domain names;
but as the conversion is applied to the whole email address, umlauts
in the local part of an email address also get converted and the email
address subsequently is regarded as valid though umlauts are not allowed
in the local part according to RFC 5321 where only normal ASCII
characters (plus some special characters) are allowed
(see http://tools.ietf.org/html/rfc5322#section-3.2.3).

To fix the bug, the email address has to be split and only the domain
part must be converted using IDNA; the email address is then composed
again and validated.

EDIT: Fix violation of TYPO3 coding guidelines.

Resolves: #44103
Releases: 6.0

Change-Id: Ia623901eaed41a73fdc2d934d58c1e621c349cd0