[BUGFIX] Impossible to save the TCA configuration
The configuration module allows to save changes to the TCA array.

If no extTablesDefinitionScript is defined, constant
TYPO3_extTableDef_script is defined as empty. This results in an
attempt to write to a directory (because the filename is empty).
This is not obvious as the constant is not needed much and the
error message is misleading.

Solution is to fall back to the default value during bootstrap if
extTablesDefinitionScript is not set in LocalConfiguration.php.

Change-Id: Id759aa73146669cccae9a83d8452340f9f78927b
Fixes: #44331
Releases: 6.0, 6.1
Reviewed-on: https://review.typo3.org/17381
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn