[BUGFIX] Impossible to save the TCA configuration
The configuration module allows to save changes to the TCA array.

If no extTablesDefinitionScript is defined, then the constant
TYPO3_extTableDef_script is defined as empty. This results in the attemp
to write to a directory (because the filename is empty).
This is not obvious as the the constant is not needed much and the error
message is very misleading.

Make the bootstrap fall back to the usual value of the
extTablesDefinitionScript is not defined.

Change-Id: Id759aa73146669cccae9a83d8452340f9f78927b
Fixes: #44331
Releases: 6.0, 6.1