[BUGFIX] Content elements in wrong column in page module
If working on workspaces, it might happen that content elements
just appear in the wrong column. The reason is, that live(!)
content elements are selected using a specific column value and
then are overlaid with workspace data.

If the column value has been modified in a workspace, this is
currently just ignored during rendering.

Change-Id: I6669ea096d93269591727b20d92ef5250898900f
Fixes: #44470
Releases: 6.1, 6.0, 4.7, 4.6, 4.5