[BUGFIX] Content elements in wrong column in page module
If working on workspaces, it might happen that content elements
just appear in the wrong column. The reason is, that live(!)
content elements are selected using a specific column value and
then are overlays by workspace data.

If the column value has been modified in a workspace, this is
currently just ignored during rendering.

Change-Id: I9eccc2ff5b7220544cd28ee6de8f55453c04a5d1
Fixes: #44470
Releases: 6.1, 6.0, 4.7, 4.6, 4.5