[BUGFIX] Allow Setting colorspace in the Install Tool.
Some versions of Imagemagick (6.7.0 and above) use the sRGB colorspace
instead RGB as before. This results in darker images after processing,
because TYPO3 hardcoded the RGB colorspace in graphical functions.

This patch introduces a setting in the GFX part of the Install Tool,
lets the user choose the sufficient colorspace.
This selection is used in graphical functions.

Additionaly a hint on the setting was added to the Image Processing Test
'Read Images' in the Install Tool.

Fixes: #36597
Releases: 6.1, 6.0, 4.7, 4.6, 4.5
Change-Id: I50a26c414705afa3177a2f12fc3bb4532c2d0f7f