[BUGFIX] Translation of a form makes the form wizard unusable
As soon as the form-configuration for the form wizard gets
loaded, the leading "[Translate to XYZ:] " is stripped off to get the
wizard working (with the leading text, the form config could not
be serialized as needed and the wizard didn't show a single field).

The stripping is only done on localized records.

Change-Id: I0cab1d371abf5ae01b524c39acedcd1a3a5db9c4
Fixes: #39127
Releases: 4.7, 6.0, 6.1
Reviewed-on: https://review.typo3.org/12996
Tested-by: Mario Rimann
Reviewed-by: Stefan Galinski
Tested-by: Stefan Galinski
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn