[BUGFIX] Remove inline styles from parsed blockquote tag
When passing blockquote tags trough lib.parseFunc_RTE (this
is commonly done when rendering TS objects in Fluid templates)
their attributes will be replaced with some inline styles.
I don't know what the intention of this behavior is, but it's been
like this at least since 2004. Inline styles should generally be
avoided.
I suggest to simple remove this rule.

Code to reproduce the issue and test the change:
https://gist.github.com/4654874

Resolves: #44879
Releases: 6.1, 6.0, 4.7
Change-Id: If8954fa392b906d78a8cc32aba9e5b23d84decd9