[BUGFIX] Possible warning in about module
ext:about includes ext_emconf.php files of loaded extensions.
The code raises a warning if those files are not found. This
can happen for example, if some extension is loaded in
LocalConfiguration "extList", but does not exist in the file
system.

Solution is to check for file existance before include.

Change-Id: Ic58adfc908ff02d5f0d9e79b69719d02c6b429ab
Fixes: #44892
Releases: 6.1, 6.0, 4.7
Reviewed-on: https://review.typo3.org/17768
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn