[TASK] EXT:reports Improve xclass reporting
Since TYPO3 CMS 6.0, the XCLASS registration changed and the old
registration style does not work anymore. Additionally, 6.0 comes
with a basic check for old XCLASS registration to warn users about
that.

The patch adds a new report for newly registered XCLASS'es and
adapts the messages to better notify that neither old nor new
registrations are a real problem, but that those classes should
be kept an eye on during upgrading.

Change-Id: I8c0052b9c7fb0e88aff62a71c9592bf51bcec7ad
Resolves: #44895
Releases: 6.1, 6.0
Reviewed-on: https://review.typo3.org/17778
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn