[BUGFIX] LocalDriver: Recursive file listing is broken
The recursive file listing, introduced as part of the public API in
the fix for #43249, is currently broken.

One problem is that the file information retrieval is incomplete: The
filename of files in subfolders also contains the path to this file from
the current folder, while the identifier is missing that information.
The identifier is thus wrong and the filename contains too much
information (more than the filename).

Additionally, the method getDirectoryItemList() returns the file list
with the filenames as key, which will fail when a file name exists twice
in different folders. Therefore, this patch changes the keys to numeric
values when a recursive folder list is requested.

Change-Id: Iaebd862327d2dfc849044236474f6da2444cd4f5
Resolves: #44910
Releases: 6.1, 6.0
Reviewed-on: https://review.typo3.org/17881
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter