[FEATURE] FLUIDTEMPLATE cObj settings array support
extbase action controllers give a settings typoscript sub array
to the view, if it exists. This feature is very handy, it allows
full typoscript magic like constants access on this sub array.

The FLUIDTEMPLATE content object has a the variables typoscript
configuration, but this allows only other content objects in it. This
is unpractical if just some key-value nested array should be given
to the view.

The patch adds a settings property to the FLUIDTEMPLATE cObj that
behaves similar to the extbase action controller. If given, this
array is just assigned as is to the view.

Example:

lib.globalSettings {
	# Access to constants works, too.
	foo = value
}
page = PAGE
page.10 = FLUIDTEMPLATE
page.10 {
	file = fileadmin/templates/MyTemplate.html
	settings < lib.globalSettings
}

In the view, the value can then be used: {settings.foo}

Change-Id: Ie48911275832d422525e7d8891096fefcc0af152
Resolves: #23853
Releases: 6.1