[WIP][FEATURE] Search within the FAL RecordPicker
Adds a search field to the File ElementBrowser, allowing
to search files after text match in file name and meta data.
The configured search fields of sys_file are taken into
account accordingly.

This is a massive rewrite of the patch Steffen Ritter originally wrote;
as of now, it is not ready for commit, but just a proof of concept.
DO NOT COMMIT THIS NOW!

Change-Id: I019e05bd5a7cb8837ce98a15e55e1eb5a6b56f74
Resolves: #43286
Releases: 6.1