[BUGFIX] Remove delete button for sys_file
The edit view of file meta data currently has
a delete button, which sets the record of this
file to deleted, but does not delete the file itself.

It might be useful at some point
to use this to implement some recycler
functionality, but currently the FAL layer
has no support for it and a file with a deleted
meta data record is an inconsistent state
which has to be avoided by all means.

Resolves: #45065
Releases: 6.0, 6.1

Change-Id: I5033107cf4021d127e274bbe12e3aa08ea1b4ef6