[BUGFIX] Flexform "required" on input fields applies to last field only
When a flexform has several input fields with "required" eval option,
only the last one is validated. The requirement for all but the last one
is ignored.

This patch fixes the bug by using the flexform field name instead of
the database field name as array key.
This does still work for "normal" fields, since both values are
populated (tested with "htmlArea RTE abbreviation").

Change-Id: I6581a7459eec81d329a06db722e3e574e7666568
Fixes: #17122
Releases: 6.0, 6.1