[BUGFIX] OpenId needs library after removing classFile
After removing the classFile setting for services the OpenID
authentication service needs to have its library loaded
explicitly. Unlike the rest of the core this external
library doesn't use namespaces and can't rely on the autoloader.
Furthermore, the library might not have a unique name if
for example a second extensions delivers the same library,
therefor the problem is for now not solved with an ext_autoload
entry, but is required directly at where needed.

Change-Id: Ibb364ebbf3a16f63f61bef3b1987963b6c7f881a
Fixes: #45095
Related: #37115
Releases: 6.1