[BUGFIX] Date ViewHelper not using configured Timezones
The date viewhelper does not respect the configured timezone.

After the default timezone is set during Bootstrap, it is enough to
retrieve the setting from environment.

Additional a condition tests the proper format of timestamps and adds
the '@' if it is missing.

For the formatted output with '%' strftime function is added, too.

Fixes: #12769, #43363, #9174
Releases: 4.5, 4.6, 4.7, 6.0, 6.1
Change-Id: I782b4ec00537519768335da9ba32822a42108ea1
Reviewed-on: https://review.typo3.org/17976
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Jigal van Hemert
Tested-by: Jigal van Hemert
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring