[BUGFIX] Faulty translation behavior of IRRE using workspaces
The translation behavior of IRRE child records on workspaces is
wrong on plain new records.

Inline Relational Records editing used the most specific version
of a record for references. In case of creating new records on a
workspace, a version placeholder record will be generated as
well as a specific version to overlay that placeholder.
Translations of a record will copy the accordant record and
refer to it by using the l18n_parent field (which is the
"transOrigPointerField" in TCA). For historic reasons the uid of
the placeholder record is used in that case instead of the most
specific version - otherwise relations between original and
translation are not recognized in several places in TYPO3 (which
is actually the real bug, but cannot be resolved easily on the
fly for historical reasons).

Thus, the bug fix is to use the uid of the placeholder record in
those cases instead and to adopt the form view component of IRRE
to recognize those placeholders as translation parents as well.

Change-Id: I0a413ead5f127e17cde03737edddf49688dcad20
Fixes: #45132
Releases: 6.1, 6.0, 4.7, 4.6, 4.5