[BUGFIX] New IRRE parent-child-structures are not removed
If creating new records on a workspace, a placeholder record
will be created and overlayed by a specific version. Removing
the parent record does not remove the dependent 1:n children
which leads to warnings on removing translations and recreating
them afterwards (since child records are not really removed, but
lost their parent reference).

The fix is to process placeholders on removing after references
have been removed. Currently it's the other way round which
leads to the situation that parents just lost their children.

Change-Id: I9682334b2e31cff025bcb1cd642bd98773c0b244
Fixes: #45133
Releases: 6.1, 6.0, 4.7, 4.6, 4.5