[BUGFIX] show_item does not consider workspace references
typo3/show_item.php does not consider workspace references, that
might be IRRE or plain MM. The reason is that the overlay still
uses the uid of the original record (which is the placeholder on
new workspace records).

The fix is to use the processed record index "_ORIG_uid" instead
of "uid" here, if available.

Change-Id: If54fe3ad235a5f33848bf88b73e2ab2cd3688fc1
Fixes: #45148
Releases: 6.1, 6.0, 4.7, 4.6, 4.5