[BUGFIX] Problems on publishing complex IRRE structure
Imagine the following IRRE parent-child structure. The important
thing is, that Price is created by different parents.

+ Hotel
\-+ Offer
| \-+ Price
\-+ Flight
  \-+ Price

On publishing this structure, most likely the last Price element
gets lost. The reason for that is simple and located in the
command map resolver, that is used to correctly publish the
single elements from outer-most parent (root) to all dependent
children (leafs). The result of the resolver is a "cmdmap" array
to be used in t3lib_TCEmain/DataHandler.

The Price object below Offer already defines the array elements
for Price which gets completed by the child of Flight later on.
The resulting "cmdmap" looks like this:

cmd[Hotel][1] = ...
cmd[Price][1] = ...
cmd[Price][2] = ...
cmd[Flight][1] = ...

However, all elements of the Price object must be process after
the Flight object, which is a relative parent in this case.

Change-Id: Ifc28b6aeb38bf5d2222d1f8e977c1e562537feb0
Fixes: #45214
Releases: 6.1, 6.0, 4.7, 4.6, 4.5