[BUGFIX] inject* methods in FAL inhibit use of Extbase object manager
The ResourceStorage class of FAL has two inject* methods which are seen
as dependency injectors by the Extbase object manager. Thus an exception
is thrown when a ResourceStorage object is created via Extbase.

This commit changes the two method's names and adjust all usages within
the core. These methods are internal and I don't expect them to be used
anywhere else, so we don't need a backwards compatibility layer.

Change-Id: I4b73445708cf8041b505b5ea8923dc595ddaadb0
Resolves: #45319
Releases: 6.1, 6.0