[TASK] Deprecate TimeTracker
The TimeTracker is one of the oldest parts of TYPO3 and
mostly unused.

During the restructuring of the ContentObjectRenderer and other parts
of the Bootstrap including the patches to the PageRenderer the
TimeTracker didn't get considered at all and still bloats the code.
The traces produced by the TimeTracker are thus incomplete. Profiling
should not be a core tasks of and the TimeTracker deprecated and
removed.

This patch removes all calls on the global TimeTracker and
replaces the global TimeTracker instance by the NullTimeTracker.

Resolves: #45329
Releases: 6.1
Change-Id: I20d4bb94fea8d575b0673a310b3f8ca59665e0a1