[BUGFIX] Exception when scheduling "at" execution time
If a recurring task is converted to a single task and the endtime is set
before now, then the exception "Task is past end date" is thrown as the
next execution time for the "at" daemon is calculated.

The fix is to catch the exceptions and ignore them, as we don't care
about tasks that will never run again when calculating the next
execution date.

Fixes: #45265
Releases: 6.1, 6.0
Change-Id: If6fd368e6c11c0b9058eb8a82a736220ef538b70
Reviewed-on: https://review.typo3.org/18106
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn