[BUGFIX] Usage of id=pageAlias is broken while logged in in BE
The page alias is not converted to the page id if you are logged in
in the BE. To fix this and correctly handle not existing aliases,
we use \TYPO3\CMS\Frontend\Page\PageRepository::getPageIdFromAlias
to determine the page's uid. If there is no id found for this alias
 - the id is zero - we must not try to get the rootline, this would
cause an exception.

Fixes: #45289
Releases: 6.2, 6.1, 6.0
Change-Id: Ia5749dd1e4dcc3c520366dd5fcfdcf9d34469ab5