[BUGFIX] Do not overwrite image caption field in content adapter
Problem:
Create a new content element of type file links, add a file relation and a description to this file.
Switch this element to be an image element, add an image and a description of the image

Expected result:
An image with the image description is shown

Actual result:
An image with the description of the file is shown.

Solution:
Check if the field actually is shown in the selected
content element type.

Fixes: #45415
Releases: 6.0, 6.1

Change-Id: I3aa0e47b336cfa881c637b7f5a9e68476ded88ae