[BUGFIX] IRRE 1n:csv parent records point to workspace placeholder
IRRE 1n:csv parent records point to workspace placeholder
instead of using the specific versioned record on copying.
If records are copied or newly created on a workspace, a
version placeholder is created and overlayed by the specific
version of the record.

The fix is to resolve the specific version uid of the record.

Change-Id: Ib962a79562c57c9a41071f053f596e6ba63859fe
Fixes: #45151
Releases: 6.1, 6.0, 4.7, 4.6, 4.5