[BUGFIX] Skip unreliable APC test on PHP versions below 5.3.4
APC has some slam protection that tries to prevent hammering of cache
entries. This can be disabled, but the option does not work at least
in native PHP 5.3.3 on debian squeeze. While it is no problem with
higher PHP version like the current one on travis-ci.org,
some tests are now just skipped on PHP environments knows for issues.

Change-Id: Ibbef201805b58041bb0b3e8aa16f0c1a2e366020
Resolves: #46267
Release: 6.1, 6.0
Reviewed-on: https://review.typo3.org/18901
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn