[BUGFIX] tslib_content_Media does not clean up $conf['file.']
When a url is passed to a MEDIA object, a proper
url is extracted. This will make sure a youtube
url is correctly rewritten. The result of the
cleaning action is written to $conf['file'].

When an url is passend in through $conf['file.']
and the url is cleaned, the result is written to
$conf['file'] but the original $conf['file.'] is
left untouched.

When the conf object is passed onto SWFOBJECT
for final rendering, the first property that is
checked is the dirty $conf['file.'] and not the
clean $conf['file'].

Currently this prevents proper rendering of a
video when using a 'file.' object to render a
video url.

The solution is to unset $conf['file.'] after
$conf['file'] has been set.

Fixes: #45426
Releases: 4.5
Change-Id: I15c7020fb4b90aa57d0a3d0530d2d5d4ac503a72