[FEATURE] Display error messages when running imagemagick commands
When the install tool searches for image magick installations, it now
displays all command line errors what are encountered when trying to
run the image magick commands, such as missing library errors and so on.
This makes it easier to get an image magick installation up and running
if the TYPO3 installer does not seem to detect image magick, where it
actually just has trouble executing the command.

Change-Id: I56b9de8e59ae3afc763b19db02081527ef678162
Releases: 6.1
Fixes: #45487