[BUGFIX] Fix of .zip exports of extension
Before this, hidden files and directories (the ones starting with a
dot) where included when exporting an extension as .zip. This is no
longer the case, with the exception of .htaccess files.

Also the function responsible for the export is cleaned up and uses
more core API then before.

Fixes: #45404
Releases: 6.0
Change-Id: I4445daeaaee54bb111c87ffeb4401ce3b7d3e97e