[BUGFIX] TCA:Suggest wizard doesn't work for select with maxitems=1
For select fields with maxitems = 1, the suggest wizard does
display, but you cannot select an item and transfer it to the
selected list. Once you change it to maxitems > 1 everything
works as expected.

Solution: Avoid maxitems = 1 (unique) select fields being rendered as
multiple selector boxes and add JS that is required whenever a suggest
wizard is rendered.

Fixes: #23789
Releases: 4.5, 4.7, 6.0, 6.1
Change-Id: I1847c9dbf2fe3b4072743857bb84b8e6b112598b