[BUGFIX] Fix processed img rendering if original contains disallowed chars
Filenames of files uploaded in TYPO3 CMS before 6.0.1 can contain
problematic characters, because filename sanitizing of added files was
just introduced with 6.0.1.
The filenames of processed files (the ones prefixed with 'csm_' or
'preview_' in the _processed_ folders) are sanitizied during their
generation, but the unsanitized original filename is used for rendering
of the processed files, which causes wrong generated filepathes and the
preview/scaled images do not show up in BE/FE.

This patch also sanitizes the filename used to render processed files.

Change-Id: I53e4eb42def291ba88ce18209a348b1e2f592185
Resolves: #45221
Related: #42925
Releases: 6.1, 6.0