[FEATURE] Only use one select to calculate page cache timeout
To get the first time value we do 2 SQL select statements at the
moment but this can be combined in one SQL select statement. This
will save as time and reduces bandwidth to/from the SQL server.

The one SQL select statement from 4.6 had a bug and was changed
to use 2 SQL select statements instead of fixing the statement itself.
So this is a revert to the old source with the fixed statement.

Resolves: #38838
Releases: 6.1

Change-Id: Ia3b106cf0232ac80d287a8c50f7c8befaf0cc29b