[TASK] Deprecate class FE_loadDBGroup
Class FE_loadDBGroup's only purpose is to set an attribute of its parent
class. This should better be done when instantiating an object.

This patch additionaly renames the mentioned attribute to make its
purpose clearer. A setter method for that attribute is added and a
compatibility layer as well.

Change-Id: I44b3e331dcb86d75f266d9e36cf0829998f1fb5b
Resolves: #45937
Related: #23977
Releases: 6.1