[BUGFIX] Initialize TCAcachedExtra without resetting TCA in BE.
When TypoScriptFrontendController-getCompressedTCarray is called in BE,
ie. to simulate FE environment and give to getRecordOverlay
the possibility to access localization information stored in
TCAcachedExtra, the global TCA array is resetted and the additonal
informations from ext_tables.php extensions are no more available.

This patch moves the code snippet responsible of loading
the TCAcachedExtras in a new method to allow to get a full
TCAcachedExtra informations without resetting TCA array.

If merged, getNewCompressedTCArray will be flagged as deprecated in 6.1

Fixes: #45959
Related: #45963
Releases: 4.5, 4.6, 4.7, 6.0

Change-Id: I2c15bdf1e896899921b3eaed99d80e899deb8348