[TASK] Make the update-from-TER link more visible
Makes the "update from TER" link in the extension manager more
visible by adding a "refresh"-icon in front of it and linking
the icon instead of the text.

Also moved the rendering functionality of that link from a viewhelper
to a partial, as partials tend to be more readable and bug free.

Resolves: #46029
Releases: 6.1
Change-Id: If447bba28bc0dee3e7de286e1a93e062ef59843c