[FEATURE][CONF]Add language switcher for backend pagetree
This patch adds a language switcher in the top panel of the
pagetree for multilanguage web sites.

New TSConfig options:
options.pageTree.languageSelection.hide=1 in the user TSConfig
will hide the language switcher for this user.
options.pageTree.languageSelection.sortAlphabetically=1 in the
user TSConfig will sort the language selection alphabetically
for this user.

Resolves: #46017
Releases: 6.1
Change-Id: Ia895e0563991dd9b572df510ae0e9cc521f3eb80