[FEATURE][CONF][BUGFIX]Add language switcher for backend pagetree
This patch adds a language switcher in the top panel of the
pagetree for multilanguage web sites.

New TSConfig options:
options.pageTree.languageSelection.hide=1 in the user TSConfig
will hide the language switcher for this user.
options.pageTree.languageSelection.sortAlphabetically=1 in the
user TSConfig will sort the language selection alphabetically
for this user.

Additonally this patch fixes some bugs:
- defining a temporary mountpoint and then searching in the
  mounted tree some pages will be shown as subpages of themselves.
- a user which is restricted to only non default languages is
  able to disable/enable and to edit the default page via
  context menu of the page tree and to edit the default page
  title directly in the pagetree.

Resolves: #46017
Releases: 6.1
Change-Id: Ia895e0563991dd9b572df510ae0e9cc521f3eb80