[BUFIX] External URL regression by jumpurl security fix
Pages of type 'Link to external URL' throw a jumpurl
hash exception if called in the frontend. This typically
happens if a HMENU renders such page links.
The patch adapts the TSFE logic to write the required
hash dynamically to _GET to make the jumpurl check happy
that is called later on within the same process if
calling such a 'external url' link.

Change-Id: Ie73cd9710929c72aad4e75543f90c8618797997b
Resolves: #46071