[BUFIX] External URL regression by jumpurl security fix
With the jumpurl security fix, pages of type
'Link to external URL' throw a jumpurl
hash exception if called in the frontend. This typically
happens if a HMENU renders such page links.

The patch adapts the TSFE logic to write the required
hash dynamically to _GET to make the jumpurl check happy
that is called later on within the same process if
calling such a 'external url' link.

Change-Id: Ie73cd9710929c72aad4e75543f90c8618797997b
Fixes: #46071
Related: #28587
Releases: 6.1, 6.0, 4.7, 4.6, 4.5
Reviewed-on: https://review.typo3.org/18767
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel