[BUGFIX] BELog module error: "vsprintf(): Too few arguments"
AbstractExceptionHandler contains a function named writeLog. That
function adds an entry to the system log. System log expects not
just a string but a string format with % characters. Callers of
this methods may supply a URL ("Request URL:" in the log message)
that contains % characters. vsprintf() in the BELog view helper
will complain about missing arguments to such string.

Change-Id: Ie080c0af1d93c28f8521b670a8ae8e63f8c4cfda
Resolves: #46077
Releases: 6.0, 6.1