[FEATURE] Introduce xcache cache backend
xcache is another PHP opcode cache like APC. It features a key-value
storage.
Make this storage available to the TYPO3 caching framework by providing
a cache backend.

This code is based in the wincache backend which itself is based on the
APC backend.
Unit test are taken from wincache too.

Resolves: #46142
Releases: 6.1
Change-Id: I4e9e856a8397978eb6f6d480b79996ff49ff8a7b