[BUGFIX] HMENU rendering uses old tslib_ class names
The menu rendering in frontend still instantiates menu classes with
its old tslib_ prefixed names. This is unflexible and a potential
performance issue.

The patch introduces a simple factory instead with a mapping for
existing TypoScript menu types to the implementing classes.
The current variable $menuclasses is obsolete.
Additionally, the naming of three existing classes is adapted to
the correct class names to be in sync with the others.

A method in the factory to register additional menu types will
be added with a second patch.

Change-Id: I3d65258ffe4d30a918705b23349d3facb563c80e
Resolves: #46292
Releases: 6.1, 6.0
Reviewed-on: https://review.typo3.org/18973
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn