[BUGFIX] Missing labels in backend modules
There seems to be some magic in SOBE that automatically includes
language files if they are located parallel to the module that is
loaded. The task from #46143 and #46296 registered new wizard classes
at a different location, so the magic doesn't hit anymore. As a
result, no labels are show in above extensions in the backend.
Fix is to include the locallang file in main() of the modules. The
direct access to $LANG is removed along the way by accessing
$GLOBALS['LANG'] directly.

Change-Id: Iecddb181141f5317b87e405ada925c61cfd9261c
Fixes: #46296
Related: #46143
Related: #46296
Releases: 6.1