[FEATURE] Hardcoded CLI arguments from $_SERVER vars
Currently there is no way to overwrite the used CLI
arguments from the class who extends from the
\TYPO3\CMS\Core\Controller\CommandLineController

This occurs because the
\TYPO3\CMS\Core\Controller\CommandLineController connects
hardcoded to the $_SERVER['argv'] array. This is in
some situations very unhandy - e.g. trigger lowlevel
cleaner during a scheduler task.

Kudos to AOE, Michael Klapper and Tolleiv Nietsch!

Resolves: #21389
Releases: 6.1
Change-Id: I42c14d37e8e7f977b3c7e73b0009e6c8495c3258
Reviewed-on: https://review.typo3.org/18929
Reviewed-by: Dmitry Dulepov
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn